+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [2.2.1] - 2021-04-19
+* Adding option for callback function to ImageLabel and MultiLabel for when the text processing is completed
+* Update MultiLabel to only display once all text have been generated
 
-## [2.2.0] - 2021-04-05
+## [2.2.0] - 2021-04-16
 * Revert changes on how fractions are created by the MultiLabel
 * Migrate math syntax to the MultiLabel